requirements.txt
# API Gateway & Todo App Requirements
# Install with: pip install -r requirements.txt

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.2

# Development
black==23.12.0
pylint==3.0.3
flake8==6.1.0

# Additional utilities
pydantic==2.5.0
pydantic-settings==2.1.0
